refactor(matcher): simplify and rename hash/key matchers
- Replaced HashMatcher with ConditionMatcher as the field matcher dispatcher.
- KeyValueMatcher is removed and replaced with DigValueMatcher, now focused solely on record key traversal.
- Moved matcher dispatch logic into ConditionMatcher.
- DigValueMatcher now inherits MainMatcher for simpler reuse of match? behavior.

This flattens the matcher hierarchy and improves clarity, separation of concerns, and traceability.
