+++ lib/mongory/matchers.rb
@@ -7,9 +7,9 @@ require_relative 'matchers/eq_matcher'
 require_relative 'matchers/exists_matcher'
 require_relative 'matchers/gt_matcher'
 require_relative 'matchers/gte_matcher'
-require_relative 'matchers/hash_matcher'
+require_relative 'matchers/condition_matcher'
 require_relative 'matchers/in_matcher'
-require_relative 'matchers/key_value_matcher'
+require_relative 'matchers/dig_value_matcher'
 require_relative 'matchers/lt_matcher'
 require_relative 'matchers/lte_matcher'
 require_relative 'matchers/main_matcher'
