+++ lib/mongory/matchers/condition_matcher.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+
+require_relative 'abstract_multi_matcher'
+
+module Mongory
+  # Temp Description
+  module Matchers
+    # Temp Description
+    class ConditionMatcher < AbstractMultiMatcher
+      def build_sub_matcher(key, value)
+        case key
+        when *Matchers::OPERATOR_TO_CLASS_MAPPING.keys
+          Matchers.lookup(key).new(value)
+        else
+          DigValueMatcher.new(key, value)
+        end
+      end
+
+      def operator
+        :all?
+      end
+    end
+  end
+end
