+++ lib/mongory/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module Mongory
-  VERSION = '1.3.2'
+  VERSION = '1.3.3'
 end
